+++ packages/core/src/platform/builtInMixins/index.js
@@ -7,7 +7,7 @@ import relationsMixin from './relationsMixin'
 import i18nMixin from './i18nMixin'
 import pageTitleMixin from './pageTitleMixin'
 import pageScrollMixin from './pageScrollMixin'
-import pageOnResizeMixin from './pageOnResizeMixin'
+import pageResizeMixin from './pageResizeMixin'
 
 export default function getBuiltInMixins (options, type) {
   let bulitInMixins = []
@@ -18,7 +18,7 @@ export default function getBuiltInMixins (options, type) {
       pageTitleMixin(type),
       pageStatusMixin(type),
       pageScrollMixin(type),
-      pageOnResizeMixin(type)
+      pageResizeMixin(type)
     ]
   } else {
     // 此为差异抹平类mixins，原生模式下也需要注入也抹平平台差异
